Scope patchers under Specialist Publisher
This makes new addition of patching schema files inside of another application (in this case Rummager) easier to make.
