+++ lib/tasks/patch/countries.rake
@@ -1,8 +1,8 @@
-require 'dfid-transition/patch/countries'
+require 'dfid-transition/patch/specialist_publisher/countries'
 
 namespace :patch do
   desc 'Patch the DFID schema with countries from the GOV.UK register'
   task :countries do
-    DfidTransition::Patch::Countries.new.run
+    DfidTransition::Patch::SpecialistPublisher::Countries.new.run
   end
 end
