+++ lib/tasks/patch/regions.rake
@@ -1,8 +1,8 @@
-require 'dfid-transition/patch/regions'
+require 'dfid-transition/patch/specialist_publisher/regions'
 
 namespace :patch do
   desc 'Patch the DFID schema with regions from the SPARQL endpoint'
   task :regions do
-    DfidTransition::Patch::Regions.new.run
+    DfidTransition::Patch::SpecialistPublisher::Regions.new.run
   end
 end
