+++ spec/fixtures/schemas/rummager/dfid_research_outputs.json
@@ -1,0 +1,19 @@
+{
+  "fields": [
+    "country",
+    "published_date"
+  ],
+
+  "expanded_search_result_fields": {
+    "country": [
+      {
+        "label": "United Kingdom",
+        "value": "UK"
+      },
+      {
+        "label": "Azerbaijan",
+        "value": "AZ"
+      }
+    ]
+  }
+}
